+++ pkg/cli/server.go
@@ -201,8 +201,7 @@ func newServerService(c *config.Server) (*serverService, error) {
 	svc.scrapeManager = scrape.NewManager(
 		svc.logger.WithField("component", "scrape-manager"),
 		ingester,
-		defaultMetricsRegistry,
-		!svc.config.RemoteWrite.Enabled)
+		defaultMetricsRegistry)
 
 	svc.controller, err = server.New(server.Config{
 		Configuration:           svc.config,