+++ pkg/scrape/scrape.go
@@ -65,11 +65,9 @@ type scrapePool struct {
 	// set of hashes.
 	activeTargets  map[uint64]*Target
 	droppedTargets []*Target
-
-	disableCumulativeMerge bool
 }
 
-func newScrapePool(cfg *config.Config, p ingestion.Ingester, logger logrus.FieldLogger, m *metrics, disableCumulativeMerge bool) (*scrapePool, error) {
+func newScrapePool(cfg *config.Config, p ingestion.Ingester, logger logrus.FieldLogger, m *metrics) (*scrapePool, error) {
 	m.pools.Inc()
 	client, err := config.NewClientFromConfig(cfg.HTTPClientConfig, cfg.JobName)
 	if err != nil {
@@ -90,7 +88,6 @@ func newScrapePool(cfg *config.Config, p ingestion.Ingester, logger logrus.Field
 
 		metrics:     m,
 		poolMetrics: m.poolMetrics(cfg.JobName),
-		disableCumulativeMerge: disableCumulativeMerge,
 	}
 
 	return &sp, nil
@@ -108,8 +105,6 @@ func (sp *scrapePool) newScrapeLoop(s *scraper, i, t time.Duration) *scrapeLoop 
 		delta:       d,
 		interval:    i,
 		timeout:     t,
-
-		disableCumulativeMerge: sp.disableCumulativeMerge,
 	}
 	x.ctx, x.cancel = context.WithCancel(sp.ctx)
 	return &x
@@ -329,8 +324,6 @@ type scrapeLoop struct {
 	delta    time.Duration
 	interval time.Duration
 	timeout  time.Duration
-
-	disableCumulativeMerge bool
 }
 
 func (sl *scrapeLoop) run() {
@@ -428,8 +421,7 @@ func (sl *scrapeLoop) scrape(startTime, endTime time.Time) error {
 	}
 
 	profile := sl.scraper.profile
-	sl.scraper.profile = profile.Push(buf.Bytes(), sl.scraper.cumulative, !sl.disableCumulativeMerge)
-
+	sl.scraper.profile = profile.Push(buf.Bytes(), sl.scraper.cumulative)
 	return sl.scraper.ingester.Ingest(ctx, &ingestion.IngestInput{
 		Profile: profile,
 		Metadata: ingestion.Metadata{
@@ -472,7 +464,6 @@ func (s *scraper) scrape(ctx context.Context, dst *bytes.Buffer) error {
 		if err != nil {
 			return err
 		}
-		req.Header.Add("Accept-Encoding", "gzip")
 		req.Header.Set("User-Agent", UserAgent)
 		s.req = req
 	}